# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Deli App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /api
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /setup:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Set up initial user
      # used as the method name of the controller
      operationId: createInitialUser
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: '#/definitions/SuccessResponse'
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authenticate:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Authenticating the user
      # used as the method name of the controller
      operationId: authenticate
      parameters: 
        - name: x-api-email
          in: header
          required: true
          type: string
        - name: x-api-password
          in: header
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: '#/definitions/AuthenticationResponse'
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Returns all the users
      # used as the method name of the controller
      operationId: usersList
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: '#/definitions/UserListResponse'
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /orders:
    # binds a127 app logic to a route
    x-swagger-router-controller: order
    get:
      description: Returns all the orders
      # used as the method name of the controller
      operationId: list
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: '#/definitions/OrderListResponse'
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  UserListResponse:
    required:
      - _id
      - name
      - email
      - role
      - admin
    properties:
      _id:
        type: string
      sid:
        type: string
      name:
        type: string
      email:
        type: string
      role:
        type: string
      admin:
        type: string
      created:
        type: string
      update:
        type: string

  OrderListResponse:
    properties:
      # The array of Orders
      id:
        type: string
      restaurant_name:
        type: string
      restaurant_contact_num:
        type: string
      customer_name:
        type: string
      customer_contact_num:
        type: string
      Address:
        type: string
      amount:
        type: string
      status:
        type: string
      created:
        type: string
      updated:
        type: string

  AuthenticationResponse:
    required:
      - token
    properties:
      token:
        type: string
      _id:
        type: string
      sid:
        type: string
      name:
        type: string
      email:
        type: string
      role:
        type: string
      admin:
        type: string
      created:
        type: string
      update:
        type: string

  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string